rm some gulp file.
